requirements.txt
flask
openai-whisper
transformers
torch
ffmpeg-python